chore: Move env > application/env & kuma-api > kuma/kuma-api
