+++ src/app/application/index.ts
@@ -11,14 +11,14 @@ import RouteView from './components/route-view/RouteView.vue'
 import { routes } from './routes'
 import can from './services/can'
 import type { Can } from './services/can'
+import type { EnvVars } from './services/env/Env'
+import Env from './services/env/Env'
 import I18n from './services/i18n/I18n'
 import type { Source } from '@/app/application/services/data-source'
 import { create, destroy, getSource, DataSourcePool } from '@/app/application/services/data-source'
 import { services as kuma } from '@/app/kuma'
 import { services as me } from '@/app/me'
 import { services as x } from '@/app/x'
-import type { EnvVars } from '@/services/env/Env'
-import Env from '@/services/env/Env'
 import type { ServiceDefinition } from '@/services/utils'
 import { token, createInjections, constant } from '@/services/utils'
 import type { Component } from 'vue'
