+++ src/app/control-planes/features.ts
@@ -1,5 +1,5 @@
 import type { Can, Features } from '@/app/application/services/can'
-import type Env from '@/services/env/Env'
+import type Env from '@/app/application/services/env/Env'
 export const features = (env: Env['var']): Features => {
   return {
     'use kubernetes': (_can: Can) => {
