+++ src/app/kuma/services/kuma-api/Api.ts
@@ -1,11 +1,11 @@
 import { RestClient } from './RestClient'
-import type Env from '@/services/env/Env'
+import type Env from '@/app/application/services/env/Env'
 
 export class Api {
   constructor(
     protected client: RestClient,
     protected env: Env['var'],
-  ) {}
+  ) { }
 
   get baseUrl() {
     return this.client.baseUrl
