+++ src/app/policies/sources.ts
@@ -2,7 +2,7 @@ import { Policy, PolicyDataplane } from './data'
 import type { PolicyType } from './data'
 import { defineSources } from '../application/services/data-source'
 import type { DataSourceResponse } from '@/app/application'
-import type KumaApi from '@/services/kuma-api/KumaApi'
+import type KumaApi from '@/app/kuma/services/kuma-api/KumaApi'
 import type { PaginatedApiListResponse as CollectionResponse } from '@/types/api.d'
 
 export type PolicyCollection = CollectionResponse<Policy>
