+++ src/app/rules/sources.ts
@@ -4,7 +4,7 @@ import {
 } from './data'
 import type { DataSourceResponse } from '@/app/application'
 import { defineSources, type Source } from '@/app/application/services/data-source'
-import type KumaApi from '@/services/kuma-api/KumaApi'
+import type KumaApi from '@/app/kuma/services/kuma-api/KumaApi'
 
 export type RuleCollectionSource = DataSourceResponse<RuleCollection>
 export const sources = (source: Source, api: KumaApi) => {
