+++ src/app/zone-egresses/sources.ts
@@ -1,7 +1,7 @@
 import { ZoneEgressOverview, ZoneEgress } from './data'
 import type { DataSourceResponse } from '@/app/application'
 import { defineSources } from '@/app/application/services/data-source'
-import type KumaApi from '@/services/kuma-api/KumaApi'
+import type KumaApi from '@/app/kuma/services/kuma-api/KumaApi'
 import type { PaginatedApiListResponse as CollectionResponse } from '@/types/api.d'
 
 export type ZoneEgressSource = DataSourceResponse<ZoneEgress>
