+++ src/services/development.ts
@@ -1,10 +1,10 @@
 import { setupWorker } from 'msw/browser'
 
+import cookied from '@/app/application/services/env/CookiedEnv'
+import type Env from '@/app/application/services/env/Env'
 import debugI18n from '@/app/application/services/i18n/DebugI18n'
 import { TOKENS as CONTROL_PLANES } from '@/app/control-planes'
 import XCopyButtonDebug from '@/app/x/components/x-copy-button/XCopyButtonDebug.vue'
-import cookied from '@/services/env/CookiedEnv'
-import type Env from '@/services/env/Env'
 import type { ServiceConfigurator, Token, TokenType } from '@/services/utils'
 import { token, get } from '@/services/utils'
 import type { FS } from '@/test-support'
